refactor: modified WinningBall with WinningNumbers
