--- src/main/java/lotto/domain/winning/WinningBall.java
@@ -1,30 +1,0 @@
-package lotto.domain.winning;
-
-import lotto.domain.number.LottoNumber;
-
-public class WinningBall {
-
-    private final LottoNumber number;
-    private final boolean isBonus;
-
-    public WinningBall(String number) {
-        this(number, false);
-    }
-
-    public WinningBall(String number, boolean isBonus) {
-        this(Integer.parseInt(number), isBonus);
-    }
-
-    public WinningBall(int number, boolean isBonus) {
-        this.number = new LottoNumber(number);
-        this.isBonus = isBonus;
-    }
-
-    public LottoNumber getNumber() {
-        return number;
-    }
-
-    public boolean isBonus() {
-        return isBonus;
-    }
-}
